PROPFIND cache added
When a propfind with Depth 1 is done on a directory and
all the items are returned they get cached as if several
depth 0 requests were performed.

So when ls an instant after starts calling stat() on each
and every file, instead of making a new PROPFIND request
it is possible to reuse the same data.

I gave the cache a very short expiration time so it won't
be too unsafe.